\name{display.uos.pal}
\alias{display.uos.pal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display University of Sheffield Colours
}
\description{
Show all the colours used by the University of Sheffield branding guidelines
}
\usage{
display.uos.pal()
}
\details{
}
\value{
A ggplot is shown with all the University branding colours, their names and positions as returned by the `uos_pal` function.
}
\references{
See University of Sheffield branding guidelines for developers https://drive.google.com/file/d/1RdF4zFec9l4Ub-_VgcgI2GWq536y3fF0/view
}
\author{
Mark Dunning
}
\examples{
display.uos.pal()
}
