\name{uos_pal}
\alias{uos_pal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
University of Sheffield Colours
}
\description{
Defines all the colours used by the University of Sheffield branding guidelines
}
\usage{
uos_pal()
}
%- maybe also 'usage' for other objects documented here.
\details{
Returns a vector of official colours for the University of Sheffield
}
\value{
A named vector of colours
}
\references{
See University of Sheffield branding guidelines for developers
}
\author{
Mark Dunning
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
uos_pal()
}
