\name{display.uos.pal}
\alias{display.uos.pal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display University of Sheffield Colours
}
\description{
Show all the colours used by the University of Sheffield branding guidelines
}
\usage{
display.uos.pal()
}
\details{
}
\value{
A ggplot is shown
}
\references{
See University of Sheffield branding guidelines for developers
}
\author{
Mark Dunning
}
\examples{
display.uos.pal()
}
